requirements.txt
# requirements.txt
pyyaml>=6.0
datasets>=2.9.0
numpy>=1.20.0
tqdm>=4.62.0
requests>=2.27.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
nltk>=3.6.0
torch>=1.10.0